introduce saving/retrieving quizzes in database

+++ src/engine/controller/QuizController.kt
@@ -1,7 +1,11 @@
 package engine.controller
 
-import engine.dto.*
+import engine.dto.AnswerRequest
+import engine.dto.QuizCreateRequest
+import engine.dto.QuizResponse
+import engine.dto.SolveQuizResponse
 import engine.exception.QuizNotFoundException
+import engine.mapper.toResponse
 import engine.service.QuizService
 import engine.util.QuizMessages
 import jakarta.validation.Valid
@@ -12,7 +16,7 @@ import org.springframework.web.bind.annotation.*
 class QuizController(private val quizService: QuizService) {
 
     @GetMapping("/{id}")
-    fun getQuiz(@PathVariable id: Int): QuizResponse {
+    fun getQuiz(@PathVariable id: Long): QuizResponse {
         val quiz = quizService.getQuizById(id) ?: throw QuizNotFoundException(id)
         return quiz.toResponse()
     }
@@ -35,7 +39,7 @@ class QuizController(private val quizService: QuizService) {
     }
 
     @PostMapping("/{id}/solve")
-    fun solveQuiz(@PathVariable id: Int, @RequestBody answerRequest: AnswerRequest): SolveQuizResponse {
+    fun solveQuiz(@PathVariable id: Long, @RequestBody answerRequest: AnswerRequest): SolveQuizResponse {
         val quiz = quizService.getQuizById(id) ?: throw QuizNotFoundException(id)
         val correct = quizService.checkAnswer(quiz, answerRequest.answer)
         val feedback = if (correct) QuizMessages.CORRECT_FEEDBACK else QuizMessages.INCORRECT_FEEDBACK
